separate volume indicators
... and other cleanup
