+++ Cargo.toml
@@ -1,7 +1,9 @@
 [package]
 name = "traquer"
-version = "0.1.0"
+version = "0.0.1"
 edition = "2021"
+description = "statistical functions library"
+license = "Apache-2.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -14,5 +16,5 @@ serde_json = "*"
 criterion = "0.5.1"
 
 [[bench]]
-name = "my_benchmark"
+name = "traquer"
 harness = false
